Working on single_licence
working on the css at single_licence level, need to hardcode a lot of
stuff which isn't very good, might want to delegate more responsibilty
to the user. Also looked at the top swirl on the front-page again to get
that looking better. It's somewhat better but needs more work.